# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Risk Management System API",
        "version": "1.0.0",
        "description": "A comprehensive Risk Management Backend System for portfolio risk assessment, scoring, and analysis capabilities for individual investors. Designed for integration with MuleSoft.\n\n## Overview\nThis API provides:\n- **Risk Tolerance Assessment** - 15-question questionnaire processing and scoring\n- **Portfolio Risk Analysis** - VaR, Sharpe Ratio, Beta, Standard Deviation calculations\n- **Concentration Risk Detection** - Single position, sector, and top-5 holdings analysis\n- **Stress Testing** - Scenario analysis with historical and hypothetical scenarios\n- **Investment Suitability** - Portfolio vs risk profile alignment checks\n- **Recommendations** - Generated investment recommendations\n\n## Authentication\nThis is a demo API with **no authentication required**.\n\n## Base URLs\n- **Production:** https://risk-management-system-f04b786dc797.herokuapp.com/api/v1\n- **Local:** http://localhost:3002/api/v1\n\n## Data Note\nThis system uses an **in-memory datastore** that resets when the server restarts. It comes pre-loaded with 50 sample investors.",
        "contact": {
            "name": "API Support",
            "email": "support@riskmanagement.demo"
        },
        "license": {
            "name": "ISC",
            "url": "https://opensource.org/licenses/ISC"
        }
    },
    "servers": [
        {
...
        },
        {
...
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "type": "object",
                        "properties": {
                            "code": {
                                "type": "string",
                                "example": "NOT_FOUND"
                            },
                            "message": {
                                "type": "string",
                                "example": "Resource not found"
                            }
                        }
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            }
        }
    },
    "paths": {
...
        "/recommendations/{recommendationId}": {
            "get": {
                "summary": "Get recommendation by ID",
                "description": "Returns a single recommendation with all details",
                "tags": [
                    "Recommendations"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "recommendationId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The recommendation ID",
                        "example": "REC-001"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recommendation found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "recommendationId": {
                                                    "type": "string"
                                                },
                                                "investorId": {
                                                    "type": "string"
                                                },
                                                "profileId": {
                                                    "type": "string"
                                                },
                                                "generatedDate": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "status": {
                                                    "type": "string"
                                                },
                                                "overallAssessment": {
                                                    "type": "object",
                                                    "properties": {
                                                        "suitabilityScore": {
                                                            "type": "number"
                                                        },
                                                        "suitabilityRating": {
                                                            "type": "string",
                                                            "enum": [
                                                                "HIGHLY_SUITABLE",
                                                                "SUITABLE",
                                                                "SUITABLE_WITH_CAVEATS",
                                                                "REVIEW_REQUIRED",
                                                                "NOT_SUITABLE"
                                                            ]
                                                        },
                                                        "summary": {
                                                            "type": "string"
                                                        }
                                                    }
                                                },
                                                "recommendations": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object",
                                                        "properties": {
                                                            "recommendationId": {
                                                                "type": "string"
                                                            },
                                                            "category": {
                                                                "type": "string",
                                                                "enum": [
                                                                    "REBALANCING",
                                                                    "DIVERSIFICATION",
                                                                    "RISK_REDUCTION",
                                                                    "INCOME"
                                                                ]
                                                            },
                                                            "priority": {
                                                                "type": "string",
                                                                "enum": [
                                                                    "HIGH",
                                                                    "MEDIUM",
                                                                    "LOW"
                                                                ]
                                                            },
                                                            "title": {
                                                                "type": "string"
                                                            },
                                                            "description": {
                                                                "type": "string"
                                                            },
                                                            "currentValue": {
                                                                "type": "number"
                                                            },
                                                            "targetValue": {
                                                                "type": "number"
                                                            }
                                                        }
                                                    }
                                                },
                                                "nextReviewDate": {
                                                    "type": "string",
                                                    "format": "date"
                                                }
                                            }
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recommendation not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/recommendations/investor/{investorId}/recommendations": {
            "get": {
                "summary": "Get recommendations for investor",
                "description": "Returns all recommendations for a specific investor",
                "tags": [
                    "Recommendations"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "investorId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The investor ID",
                        "example": "INV-001"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recommendations retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "recommendations": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "count": {
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/recommendations/investor/{investorId}/recommendations/active": {
            "get": {
                "summary": "Get active recommendations for investor",
                "description": "Returns only active (non-dismissed, non-implemented) recommendations for an investor",
                "tags": [
                    "Recommendations"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "investorId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The investor ID",
                        "example": "INV-001"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Active recommendations retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "recommendations": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "count": {
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/recommendations/{recommendationId}/status": {
            "put": {
                "summary": "Update recommendation status",
                "description": "Updates the status of a recommendation (acknowledge, implement, dismiss)",
                "tags": [
                    "Recommendations"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "recommendationId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The recommendation ID",
                        "example": "REC-001"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "ACTIVE",
                                            "ACKNOWLEDGED",
                                            "IMPLEMENTED",
                                            "DISMISSED"
                                        ],
                                        "example": "ACKNOWLEDGED"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Client acknowledged during quarterly review"
                                    }
                                }
                            },
                            "example": {
                                "status": "ACKNOWLEDGED",
                                "notes": "Client acknowledged during quarterly review"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "recommendationId": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "string"
                                                },
                                                "notes": {
                                                    "type": "string"
                                                }
                                            }
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Status is required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recommendation not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}